+++ RELEASES.md
@@ -1,8 +1,12 @@
 ## 0.0.2-SNAPSHOT/20220312-14:03:32
 
+DONE:
 - Initial release
 - configuration defaults taken from ebenchDps/dps5020.ini
 - ModbusInstrument lazy constructor
+- Menu command 'scan_ports' added
+
+
 
 ## 0.0.0/18.00.2013